:recycle: fix sonar issues on Strategies/AverageStrategy.java

+++ src/main/java/hierarchical_delegation/strategies/AverageStrategy.java
@@ -6,6 +6,6 @@ import java.util.Arrays;
 public class AverageStrategy implements Strategy {
     @Override
     public ArrayList<Double> executeOperation(ArrayList<Double> data) {
-        return new ArrayList<>(Arrays.asList((double) ((double) Math.round((data.stream().mapToDouble(element -> element).sum()/data.size()) * 100)/100)));
+        return new ArrayList<>(Arrays.asList(((double) Math.round((data.stream().mapToDouble(element -> element).sum()/data.size()) * 100)/100)));
     }
 }
